datalake: renamed parquet_writer.h to arrow_writer.h
Content of the file formerly named `parquet_writer.h` doesn't contain
any parquet references
