+++ src/v/datalake/arrow_writer.h
